Hoist SwiftBinaryContainer up to a more generic NativeBinaryContainer
