+++ subprojects/language-native/src/main/java/org/gradle/language/swift/internal/DefaultSwiftComponent.java
@@ -23,6 +23,7 @@ import org.gradle.api.internal.file.FileOperations;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.provider.Property;
 import org.gradle.internal.Cast;
+import org.gradle.language.internal.DefaultNativeBinaryContainer;
 import org.gradle.language.nativeplatform.internal.DefaultNativeComponent;
 import org.gradle.language.nativeplatform.internal.Names;
 import org.gradle.language.swift.SwiftBinary;
@@ -31,7 +32,7 @@ import org.gradle.language.swift.SwiftComponent;
 import java.util.Collections;
 
 public abstract class DefaultSwiftComponent extends DefaultNativeComponent implements SwiftComponent {
-    private final DefaultSwiftBinaryContainer<SwiftBinary> binaries;
+    private final DefaultNativeBinaryContainer<SwiftBinary> binaries;
     private final FileCollection swiftSource;
     private final Property<String> module;
     private final String name;
@@ -48,7 +49,7 @@ public abstract class DefaultSwiftComponent extends DefaultNativeComponent imple
         implementation = configurations.maybeCreate(names.withSuffix("implementation"));
         implementation.setCanBeConsumed(false);
         implementation.setCanBeResolved(false);
-        binaries = Cast.uncheckedCast(objectFactory.newInstance(DefaultSwiftBinaryContainer.class, SwiftBinary.class));
+        binaries = Cast.uncheckedCast(objectFactory.newInstance(DefaultNativeBinaryContainer.class, SwiftBinary.class));
     }
 
     @Override
@@ -76,7 +77,7 @@ public abstract class DefaultSwiftComponent extends DefaultNativeComponent imple
     }
 
     @Override
-    public DefaultSwiftBinaryContainer<SwiftBinary> getBinaries() {
+    public DefaultNativeBinaryContainer<SwiftBinary> getBinaries() {
         return binaries;
     }
 }
